Consolidate `Doorkeeper::OAuth::Client::Credentials`
No need to have a separate `Methods` module IMHO. Consolidating down to
one file with everything in it. Moved tests as well.
