+++ lib/doorkeeper/oauth/client.rb
@@ -1,4 +1,3 @@
-require 'doorkeeper/oauth/client/methods'
 require 'doorkeeper/oauth/client/credentials'
 
 module Doorkeeper
