+++ spec/lib/oauth/client/credentials_spec.rb
@@ -4,6 +4,9 @@ require 'doorkeeper/oauth/client'
 
 class Doorkeeper::OAuth::Client
   describe Credentials do
+    let(:client_id) { 'some-uid' }
+    let(:client_secret) { 'some-secret' }
+
     it 'is blank when any of the credentials is blank' do
       expect(Credentials.new(nil, 'something')).to be_blank
       expect(Credentials.new('something', nil)).to be_blank
@@ -41,6 +44,44 @@ class Doorkeeper::OAuth::Client
         credentials = Credentials.from_request request, method
         expect(credentials.uid).to    eq('uid')
         expect(credentials.secret).to eq('secret')
+      end
+    end
+
+    describe :from_params do
+      it 'returns credentials from parameters when Authorization header is not available' do
+        request     = double parameters: { client_id: client_id, client_secret: client_secret }
+        uid, secret = Credentials.from_params(request)
+
+        expect(uid).to    eq('some-uid')
+        expect(secret).to eq('some-secret')
+      end
+
+      it 'is blank when there are no credentials' do
+        request     = double parameters: {}
+        uid, secret = Credentials.from_params(request)
+
+        expect(uid).to    be_blank
+        expect(secret).to be_blank
+      end
+    end
+
+    describe :from_basic do
+      let(:credentials) { Base64.encode64("#{client_id}:#{client_secret}") }
+
+      it 'decodes the credentials' do
+        request     = double authorization: "Basic #{credentials}"
+        uid, secret = Credentials.from_basic(request)
+
+        expect(uid).to    eq('some-uid')
+        expect(secret).to eq('some-secret')
+      end
+
+      it 'is blank if Authorization is not Basic' do
+        request     = double authorization: "#{credentials}"
+        uid, secret = Credentials.from_basic(request)
+
+        expect(uid).to    be_blank
+        expect(secret).to be_blank
       end
     end
   end
